+++ Assets/Scripts/GameTimer/HUDGameTimerDisplay.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ecd918346dc83f74fa6208fb7d1e54f1
+guid: 4951dcc9f165d3e48a13dbbc94bcc0e5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
