--- Assets/Scripts/GameTimer/HudTimerDisplay.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4951dcc9f165d3e48a13dbbc94bcc0e5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
